Add Twitter image and large summary card support
This adds support for recognizing a new page variable, twitter_image,
which should contain the FULL URL of the image for the webpage, and can
be defined in the page/post Front Matter.

Also, this supports a new boolean page variable, twitter_large_card, which
can also be set in the page/post Front Matter (twitter_large_card: true)
which will cause Twitter & other embedding applications to present the
preview card with a large image above the summary, rather than a small
image to the left.